Make it viable for our use-case.
It's Python, so use README.rst instead of README.md.

Turn it in to a usable script.

Document the script with arguments and a --help.

Use the standard LICENSE.txt name for the license.

Perform some cleanup in the script. Raw strings are useful for regular expressions.

Fix the contraction issue with the script, as well as with
stopwords-as-word-prefixes issue.

Since hard-wrapped text was broken, make a soft-wrap mode that behaves
better (particularly for lyrics).